fix: Make createImageBitmap detection configurable
